test: extract MockRunner
