+++ internal/app/shared_test.go
@@ -2,10 +2,6 @@ package app_test
 
 import (
 	"github.com/stretchr/testify/mock"
-
-	commanddomain "gomander/internal/command/domain"
-	"gomander/internal/event"
-	"gomander/internal/eventbus"
 )
 
 type MockLogger struct {
@@ -22,44 +18,4 @@ func (m *MockLogger) Debug(message string) {
 
 func (m *MockLogger) Error(message string) {
 	m.Called(message)
-}
-
-type MockEventEmitter struct {
-	mock.Mock
-}
-
-func (m *MockEventEmitter) EmitEvent(event event.Event, payload interface{}) {
-	m.Called(event, payload)
-}
-
-type MockRunner struct {
-	mock.Mock
-}
-
-func (m *MockRunner) RunCommand(command *commanddomain.Command, environmentPaths []string, baseWorkingDirectory string) error {
-	args := m.Called(command, environmentPaths, baseWorkingDirectory)
-	return args.Error(0)
-}
-
-func (m *MockRunner) StopRunningCommand(id string) error {
-	args := m.Called(id)
-	return args.Error(0)
-}
-
-func (m *MockRunner) StopAllRunningCommands() []error {
-	args := m.Called()
-	return args.Get(0).([]error)
-}
-
-type MockEventBus struct {
-	mock.Mock
-}
-
-func (m *MockEventBus) RegisterHandler(handler eventbus.EventHandler) {
-	m.Called(handler)
-}
-
-func (m *MockEventBus) PublishSync(e eventbus.Event) []error {
-	args := m.Called(e)
-	return args.Get(0).([]error)
 }
