--- internal/config/application/usecases/shared_test.go
@@ -1,1 +1,0 @@
-package usecases_test
